My First Machine Learning Project
This project was part of a course I took during my undergrad in machine learning. The purpose of this project was to generate different classification models to a data set of our choosing. My partner and I chose a data set consisting of optical satellite and radar measurements on agricultural crops in Canada.